Revert "Beginning of warning cleanup"
This reverts commit 2f2923637bb0a65076a7ddf455d04f6e7d2cd9e5.

+++ RSCDLCEnabler/Mods/GuitarSpeak.cpp
@@ -9,7 +9,7 @@ byte GuitarSpeak::GetCurrentNote() {
 	uintptr_t noteAdr = MemUtil::FindDMAAddy(Offsets::ptr_guitarSpeak, Offsets::ptr_guitarSpeakOffets); // Copied & Modified from GetCurrentMenu()
 
 	if (!noteAdr)
-		return (BYTE)-1;
+		return -1;
 
 	return *(byte*)noteAdr;
 }
